update rakefile and reload func

+++ Rakefile
@@ -1,29 +1,35 @@
 require_relative 'config/environment'
 $PID_FILE='dnservice.pid'
+$RKEY_FILE='dnservice.rkey'
+
+def reload_key
+	CONFIG['reload-key'] || ENV['DNS_RELOAD_KEY'] || SecureRandom.hex(64)
+end
 
 task :environment do
 	require 'securerandom'
 	sqlFile = File.join 'db', 'dnservice.sqlite3'
-	rKeyFile = 'dnservice.rkey'
-	rKey = `cat #{rKeyFile} 2> /dev/null`
+	rKey = `cat #{$RKEY_FILE} 2> /dev/null`
 	if rKey.empty?
-		ENV['DNS_RELOAD_KEY'] = ENV['DNS_RELOAD_KEY'] || SecureRandom.hex(64)
+		ENV['DNS_RELOAD_KEY'] = reload_key
 	else
+		rKey = CONFIG['reload-key'] if !CONFIG['reload-key'].blank?
 		ENV['DNS_RELOAD_KEY'] = rKey
 	end
-	File.write rKeyFile, ENV['DNS_RELOAD_KEY']
 	File.copy_stream File.join('assets', 'template.sqlite3'), sqlFile unless File.exist? sqlFile
 	ENV['DNS_DATABASE_URL'] = ENV['DNS_DATABASE_URL'] || "sqlite3://#{ File.join(File.expand_path('..', __FILE__), sqlFile)}"
 end
 
 desc 'DNService | Run Application (Not Daemon)'
 task run: :environment do
+	File.write $RKEY_FILE, ENV['DNS_RELOAD_KEY']
 	puts 'DNService Starting...'
 	`ruby dnservice.rb`
 end
 
 desc 'DNService | Start Service'
 task start: :environment do
+	File.write $RKEY_FILE, ENV['DNS_RELOAD_KEY']
 	pid = `cat #{$PID_FILE} 2> /dev/null`
 	if pid.empty?
 		puts 'DNService Starting...'
@@ -45,9 +51,10 @@ task stop: :environment do
 		puts 'DNService Exiting...'
 		`kill -QUIT #{pid} 2> /dev/null`
 		`while ps -p #{pid} > /dev/null; do sleep 1; done`
-		File.delete $PID_FILE
 		puts 'DNService Exited!!'
 	end
+	File.delete $PID_FILE if File.exist? $PID_FILE
+	File.delete $RKEY_FILE if File.exist? $RKEY_FILE
 end
 
 desc 'DNService | Restart Service'
@@ -64,14 +71,14 @@ end
 
 desc 'DNService | Reload Record'
 task reload: :environment do
+	File.write $RKEY_FILE, ENV['DNS_RELOAD_KEY']
 	Resolv::DNS.open(nameserver_port: [['127.0.0.1', CONFIG['bind-port']]]).
-		getresources(CONFIG['reload-key'], Resolv::DNS::Resource::IN::TXT)
+		getresources(reload_key, Resolv::DNS::Resource::IN::TXT)
 
 end
 
 desc 'DNService | Reset'
 task reset: :environment do
 	Rake::Task['stop'].invoke
-	File.delete 'dnservice.rkey'
 	File.delete 'log/DNService.log'
 end
